requirements.txt
tensorflow-cpu==2.4.1
pandas==1.3.0
scikit-learn==0.24.2
matplotlib==3.4.2
statsmodels==0.12.2
tensorflow-probability==0.12.0
openpyxl==3.0.7